reverted some int->float changes because of backwards compatibility for custom elements

+++ Baselet/src/com/baselet/element/OldGridElement.java
@@ -372,12 +372,12 @@ public abstract class OldGridElement extends JComponent implements GridElement {
 
 	public abstract void paintEntity(Graphics g);
 
-	protected final float textHeight() {
-		return (this.getHandler().getFontHandler().getFontSize(false) + this.getHandler().getFontHandler().getDistanceBetweenTexts(false));
+	protected final int textHeight() {
+		return (int) (this.getHandler().getFontHandler().getFontSize(false) + this.getHandler().getFontHandler().getDistanceBetweenTexts(false));
 	}
 
-	protected final float textWidth(String text, boolean applyZoom) {
-		return getHandler().getFontHandler().getTextSize(text, applyZoom).getWidth() + (int) getHandler().getFontHandler().getDistanceBetweenTexts(applyZoom);
+	protected final int textWidth(String text, boolean applyZoom) {
+		return (int) (getHandler().getFontHandler().getTextSize(text, applyZoom).getWidth() + (int) getHandler().getFontHandler().getDistanceBetweenTexts(applyZoom));
 	}
 
 	@Override
